Rename filtering module to field_filter
